refactor: 팩토리를 추상화 및 분리
테스트할때 Mock 객체 넣어주기 위함
