+++ Discussion-Korea/Presentation/Scenes/ChatRoom/ItemViewModel/ChatItemViewModelFactory.swift
@@ -1,0 +1,10 @@
+//
+//  ChatItemViewModelFactory.swift
+//  Discussion-Korea
+//
+//  Created by 이청수 on 2022/09/05.
+//
+
+protocol ChatItemViewModelFactory {
+    func create(prevChat: Chat?, chat: Chat, isEditing: Bool) -> ChatItemViewModel
+}
